com.CapVentis.ZoomableCirclePacking
Releasing this implementation of the d3 zoomable Circle Packing
(http://bl.ocks.org/mbostock/7607535) for Qlik Sense.
This has a nice effect where, when we zoom, it makes the  partial
selection.
However, if we zoom down again there is an issue where multiple
selections can be made - probably because of a change in the underlying
dataset.